location_voiture/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
